ai/requirements.txt
flask==3.1.1
flask-cors==6.0.1
sentence-transformers
transformers
torch
torchvision
ftfy
pdfplumber
python-docx
matplotlib
pillow
requests
